cambiando int to float

+++ ApotheGSF/Clases/MedicamentosDetalle.cs
@@ -7,7 +7,7 @@
         public string NombreMedicamento { get; set; }
         public int TipoCantidad { get; set; }
         public int Cantidad { get; set; }
-        public int Precio { get; set; }
+        public float Precio { get; set; }
         public float Total { get; set; }
         public bool Abierto { get; set; }
         public int CantidadAbierto { get; set; }
